2 rev with int

+++ 0x06-pointers_arrays_strings/2-strncpy.c
@@ -10,7 +10,7 @@
 
 char *_strncpy(char *dest, char *src, int n)
 {
-	int i = 0, j;
+	long i = 0, j;
 	char *c = dest;
 
 	for (j = 0 ; src[j] != '\0' ; j++)
